# Add microbenchmark and improve performance of lambda event handler