--- a/src/testJars/java/net/minecraftforge/eventbus/benchmarks/compiled/CancelableEvent.java
+++ b/src/testJars/java/net/minecraftforge/eventbus/benchmarks/compiled/CancelableEvent.java
@@ -1,0 +1,13 @@
+package net.minecraftforge.eventbus.benchmarks.compiled;
+
+
+import net.minecraftforge.eventbus.api.Cancelable;
+import net.minecraftforge.eventbus.api.Event;
+
+import java.util.function.Supplier;
+
+@Cancelable
+public class CancelableEvent extends Event
+{
+    public static Supplier<Event> makeNew = CancelableEvent::new;
+}