--- a/src/testJars/java/net/minecraftforge/eventbus/benchmarks/compiled/ResultEvent.java
+++ b/src/testJars/java/net/minecraftforge/eventbus/benchmarks/compiled/ResultEvent.java
@@ -1,0 +1,12 @@
+package net.minecraftforge.eventbus.benchmarks.compiled;
+
+
+import net.minecraftforge.eventbus.api.Event;
+
+import java.util.function.Supplier;
+
+@Event.HasResult
+public class ResultEvent extends Event
+{
+    public static Supplier<Event> makeNew = ResultEvent::new;
+}