--- a/src/testJars/java/net/minecraftforge/eventbus/benchmarks/compiled/SubscriberDynamic.java
+++ b/src/testJars/java/net/minecraftforge/eventbus/benchmarks/compiled/SubscriberDynamic.java
@@ -1,0 +1,26 @@
+package net.minecraftforge.eventbus.benchmarks.compiled;
+
+
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+public class SubscriberDynamic
+{
+
+    @SubscribeEvent
+    public static void onCancelableEvent(CancelableEvent event)
+    {
+
+    }
+
+    @SubscribeEvent
+    public static void onResultEvent(ResultEvent event)
+    {
+
+    }
+
+    @SubscribeEvent
+    public static void onSimpleEvent(SimpleEvent event)
+    {
+
+    }
+}