This adds some microbenchmarks, similar to modlauncher.
The changes in EventBus should make the code do the same, but much faster.
Before:
```
Benchmark                             Mode  Cnt  Score   Error  Units
EventBusBenchmark.testDynamic         avgt    9  0,009 ± 0,002  us/op
EventBusBenchmark.testDynamic:·stack  avgt         NaN            ---
EventBusBenchmark.testLambda          avgt    9  0,246 ± 0,008  us/op
EventBusBenchmark.testLambda:·stack   avgt         NaN            ---
EventBusBenchmark.testStatic          avgt    9  0,035 ± 0,001  us/op
EventBusBenchmark.testStatic:·stack   avgt         NaN            ---
```
After:
```
Benchmark                             Mode  Cnt  Score    Error  Units
EventBusBenchmark.testDynamic         avgt    9  0,009 ±  0,001  us/op
EventBusBenchmark.testDynamic:·stack  avgt         NaN             ---
EventBusBenchmark.testLambda          avgt    9  0,037 ±  0,002  us/op
EventBusBenchmark.testLambda:·stack   avgt         NaN             ---
EventBusBenchmark.testStatic          avgt    9  0,032 ±  0,001  us/op
EventBusBenchmark.testStatic:·stack   avgt         NaN             ---
```

Stack after:
```
 89,7%  89,7% net.minecraftforge.eventbus.ASMEventHandler.invoke
  5,9%   5,9% net.minecraftforge.eventbus.benchmarks.generated.EventBusBenchmark_testStatic_jmhTest.testStatic_avgt_jmhStub
  3,1%   3,1% net.minecraftforge.eventbus.benchmarks.EventBusBenchmark.testStatic
  0,3%   0,3% net.minecraftforge.eventbus.benchmarks.compiled.SimpleEvent$$Lambda$35/1651988869.get
  0,2%   0,2% net.minecraftforge.eventbus.benchmarks.compiled.ResultEvent$$Lambda$34/141967142.get
  0,2%   0,2% net.minecraftforge.eventbus.benchmarks.compiled.CancelableEvent$$Lambda$33/526889128.get
  0,2%   0,2% net.minecraftforge.eventbus.benchmarks.compiled.ResultEvent$$Lambda$34/508167559.get
  0,1%   0,1% net.minecraftforge.eventbus.benchmarks.compiled.CancelableEvent$$Lambda$33/1749476209.get
  0,1%   0,1% net.minecraftforge.eventbus.benchmarks.compiled.SimpleEvent$$Lambda$35/741590357.get
  0,1%   0,1% net.minecraftforge.eventbus.benchmarks.compiled.SimpleEvent$$Lambda$35/1922543470.get
  0,2%   0,2% <other>
```
dynamic is still much faster than all the others, but I don't see any way to improve this right now