@mezz 2019-04-28
I ran the benchmark on my computer and got these results:
```
# JDK 1.8.0_211, Java HotSpot(TM) 64-Bit Server VM, 25.211-b12
# i7-7700K CPU @ 4.20GHz
```
```
Benchmark                             Mode  Cnt   Score   Error  Units
EventBusBenchmark.testDynamic         avgt    9  28.004 ± 0.275  ns/op
EventBusBenchmark.testDynamic:·stack  avgt          NaN            ---
EventBusBenchmark.testLambda          avgt    9  28.270 ± 0.154  ns/op
EventBusBenchmark.testLambda:·stack   avgt          NaN            ---
EventBusBenchmark.testStatic          avgt    9  26.412 ± 0.854  ns/op
EventBusBenchmark.testStatic:·stack   avgt          NaN            ---
```

Seems well-optimized across the board to me, dynamic is not showing as significantly faster on my machine.