@ichttt 2019-04-28
Yeah it was a mistake in my benchmark, I did not register the dynamic events handler correctly, cpw fixed it in dcf30f7ad17e4122a2b0827a809fec4a89ffbfcb